% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_viz.R
\name{cforest_viz}
\alias{cforest_viz}
\title{Visualization of the Variables of Importance for cforest}
\usage{
cforest_viz(obj, data, outcome, n = 10)
}
\arguments{
\item{obj}{a randomForest object}

\item{data}{the data.frame object from which the randomForest object was made}

\item{outcome}{the name of the outcome variable in quotes (e.g., "bmi")}

\item{n}{the number of important variables}
}
\description{
Visualizes--using ggplot2--the relationship of n important variables 
in a random forest with the outcome with a loess curve (if continuous). 
You can easily add more ggplot2 functions to the plot to adjust it as
desired. The default theme is this package's `theme_anteo`.
}

