% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_grplasso.R
\name{cv_grplasso}
\alias{cv_grplasso}
\title{Cross-Validation of Group LASSO}
\usage{
cv_grplasso(formula, .data, lambda, model = LinReg(), ..., plot = TRUE)
}
\arguments{
\item{formula}{a two-sided formula for the model fit}

\item{.data}{the \code{data.frame} object}

\item{lambda}{the number(s) to test for the tuning parameter $\lambda$. Can be a sequence but must be in reverse order.}

\item{model}{default is \code{LinReg()} for linear regression, other option is \code{LogReg()} for logistic regression.}

\item{...}{other arguments for the \code{grplasso} function such as \code{model = LinReg()}}

\item{plot}{logical; when set to \code{TRUE} it prints out a ggplot2 plot of the RMSE of each lambda value supplied}
}
\description{
Performs 10-fold cross-validation for models fit with \code{grplasso}.
}

