% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_viz.R
\name{forest_viz}
\alias{forest_viz}
\title{Visualization of the Variables of Importance}
\usage{
forest_viz(obj, n = 10, imp_measure = "default")
}
\arguments{
\item{obj}{a randomForest object}

\item{n}{the number of important variables}

\item{imp_measure}{the variable importance measure}
}
\description{
Visualizes--using ggplot2--the relationship of n important variables 
in a random forest with the outcome with a loess curve (if continuous). 
You can easily add more ggplot2 functions to the plot to adjust it as
desired. The default theme is this package's `theme_anteo`.
}

