% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_extract.R
\name{lasso_extract}
\alias{lasso_extract}
\title{Extracting the Selected Variables}
\usage{
lasso_extract(..., s = "lambda.1se")
}
\arguments{
\item{...}{cv.glmnet objects}

\item{s}{options of lambda cut off (see \code{coef.glmnet})}
}
\description{
Extracts the list of selected variables from \code{cv.glmnet} lasso regression.
Can easily be combined with the data to subset with only those variables.
}

