% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_viz.R
\name{obtain_important}
\alias{obtain_important}
\title{Obtain Important Variable Information}
\usage{
obtain_important(obj, n = 10, imp_measure = "mse")
}
\arguments{
\item{obj}{a randomForest object}

\item{n}{the number of important variables}

\item{imp_measure}{the variable importance measure}
}
\description{
Internal tool for forest_viz and forest_tab1 functions. Can be used to get data
from a Random Forest object.
}

