% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_alpha.R
\name{select_alpha}
\alias{select_alpha}
\title{Selector of Alpha in Elastic Net}
\usage{
select_alpha(x, y, alpha, ...)
}
\arguments{
\item{x}{predictor matrix}

\item{y}{response vector}

\item{alpha}{a single value or a sequence of values $0 \leq \alpha \leq 1$}

\item{...}{optional arguments passed to \code{cv.glmnet}}
}
\description{
Reports the minimum cross-validated error at various, user specific alphas
for elastic net. Uses \code{cv.glmnet} from the \code{glmnet} package
}

