% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_grplasso.R
\name{selected_grplasso}
\alias{selected_grplasso}
\title{Variable Selection of Group LASSO}
\usage{
selected_grplasso(grplasso_obj)
}
\arguments{
\item{grplasso_obj}{the model object from \code{grplasso::grplasso}}
}
\description{
Provided a model fit with \code{grplasso::grplasso}, the function provides a plot
showing when each variable was selected by the value of lambda (the tuning parameter).
}

