% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_anteo.R
\name{theme_anteo}
\alias{theme_anteo}
\title{ggplot2 anteo theme}
\usage{
theme_anteo(base_size = 12, base_family = "Helvetica")
}
\arguments{
\item{base_size}{default 12}

\item{base_family}{default Helvetica}
}
\description{
A clean ggplot2 theme designed for prresentations, websites, and reports. For more 
information on ggplot2 themes, see the documentation for ggplot2.
}

