% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_anteo.R
\name{theme_anteo_wh}
\alias{theme_anteo_wh}
\title{ggplot2 anteo-white theme}
\usage{
theme_anteo_wh(base_size = 12, base_family = "Helvetica")
}
\arguments{
\item{base_size}{default 12}

\item{base_family}{default Helvetica}
}
\description{
A clean ggplot2 theme designed for prresentations, websites, and reports similar to
theme_anteo() but with a white background for better printing. For more 
information on ggplot2 themes, see the documentation for ggplot2.
}

